Move ProcessUtils.execAndCheckOutput to ExecUtils.spawnAndGetOutput
to avoid circular dependency from IO -> PocessUtils -> Logger -> IO
