+++ packages/imperative/src/plugins/utilities/npm-interface/uninstall.ts
@@ -16,7 +16,7 @@ import { readFileSync, writeFileSync } from "jsonfile";
 import { IPluginJson } from "../../doc/IPluginJson";
 import { Logger } from "../../../../../logger";
 import { ImperativeError } from "../../../../../error";
-import { ProcessUtils, TextUtils } from "../../../../../utilities";
+import { ExecUtils, TextUtils } from "../../../../../utilities";
 import { StdioOptions } from "child_process";
 import { findNpmOnPath } from "../NpmFunctions";
 const npmCmd = findNpmOnPath();
@@ -64,7 +64,7 @@ export function uninstall(packageName: string): void {
         // formatting or colors but at least I can get the output of stdout right. (comment from install handler)
         iConsole.info("Uninstalling package...this may take some time.");
 
-        ProcessUtils.execAndCheckOutput(npmCmd,
+        ExecUtils.spawnAndGetOutput(npmCmd,
             [
                 "uninstall",
                 npmPackage,
